Fixed bugs with img upload and edit

+++ pages/post/post_delete.php
@@ -8,6 +8,15 @@ if (!isset($_REQUEST))
 if (empty($_REQUEST['post']))
     exit;
 $postid = $_REQUEST['post'];
-$cfg->db->deletePost($postid);
 
+$result = $cfg->db->getPost($postid);
+if (empty($result) || count($result) == 0) {
+    exit;
+}
+
+if (is_file(ROOT.$result[0]['postimg'])) {
+    unlink(ROOT.$result[0]['postimg']);
+}
+
+$cfg->db->deletePost($postid);
 ?>
